add submit_asset_runs utility
